Renamed `api` package to `element`
