+++ src/main/java/com/github/simonharmonicminor/beholder/core/element/constructor/ConstructorElement.java
@@ -1,7 +1,7 @@
-package com.github.simonharmonicminor.beholder.core.api.constructor;
+package com.github.simonharmonicminor.beholder.core.element.constructor;
 
-import com.github.simonharmonicminor.beholder.core.api.param.ParamElement;
-import com.github.simonharmonicminor.beholder.core.api.Element;
+import com.github.simonharmonicminor.beholder.core.element.param.ParamElement;
+import com.github.simonharmonicminor.beholder.core.element.Element;
 import com.github.simonharmonicminor.juu.collection.immutable.ImmutableList;
 
 /**
