+++ src/main/java/com/github/simonharmonicminor/beholder/core/element/member/EnumValueElement.java
@@ -1,6 +1,6 @@
-package com.github.simonharmonicminor.beholder.core.api.member;
+package com.github.simonharmonicminor.beholder.core.element.member;
 
-import com.github.simonharmonicminor.beholder.core.api.Element;
+import com.github.simonharmonicminor.beholder.core.element.Element;
 
 /**
  * Represents a value containing in one enum
