+++ src/main/java/com/github/simonharmonicminor/beholder/core/element/method/ProcedureElement.java
@@ -1,4 +1,4 @@
-package com.github.simonharmonicminor.beholder.core.api.method;
+package com.github.simonharmonicminor.beholder.core.element.method;
 
 /**
  * Represents procedure declared in one type.
